+++ app/src/main/java/app/passwordstore/ui/pgp/PGPKeyListActivity.kt
@@ -34,6 +34,7 @@ import app.passwordstore.crypto.PGPKeyManager
 import app.passwordstore.data.crypto.CryptoRepository
 import app.passwordstore.ui.APSAppBar
 import app.passwordstore.ui.compose.theme.APSTheme
+import app.passwordstore.ui.dialogs.AddPgpKeyBottomSheet
 import app.passwordstore.ui.dialogs.PasswordDialog
 import app.passwordstore.util.extensions.snackbar
 import app.passwordstore.util.viewmodel.PGPKeyListViewModel
@@ -48,7 +49,6 @@ import java.io.IOException
 import java.security.SecureRandom
 import javax.inject.Inject
 import kotlinx.coroutines.launch
-import kotlinx.coroutines.runBlocking
 import logcat.LogPriority.ERROR
 import logcat.asLog
 import logcat.logcat
@@ -63,7 +63,8 @@ class PGPKeyListActivity : AppCompatActivity() {
   private var retries = 0
 
   private val viewModel: PGPKeyListViewModel by viewModels()
-  private val keyImportAction =
+
+  private val keyAction =
     registerForActivityResult(StartActivityForResult()) {
       if (it.resultCode == RESULT_OK) {
         viewModel.updateKeySet()
@@ -84,6 +85,12 @@ class PGPKeyListActivity : AppCompatActivity() {
   override fun onCreate(savedInstanceState: Bundle?) {
     super.onCreate(savedInstanceState)
     val isSelecting = intent.extras?.getBoolean(EXTRA_KEY_SELECTION) ?: false
+    supportFragmentManager.setFragmentResultListener(PGP_KEY_ADD_REQUEST_KEY, this) { _, bundle ->
+      when (bundle.getString(ACTION_KEY)) {
+        ACTION_IMPORT_FILE -> keyAction.launch(Intent(this, PGPKeyImportActivity::class.java))
+        ACTION_NEW_PGP_KEY -> keyAction.launch(Intent(this, PGPKeyCreationActivity::class.java))
+      }
+    }
     setContent {
       APSTheme {
         Scaffold(
@@ -99,7 +106,9 @@ class PGPKeyListActivity : AppCompatActivity() {
           },
           floatingActionButton = {
             FloatingActionButton(
-              onClick = { keyImportAction.launch(Intent(this, PGPKeyImportActivity::class.java)) }
+              onClick = {
+                AddPgpKeyBottomSheet().show(supportFragmentManager, "ADD_PGP_KEY_BOTTOM_SHEET")
+              }
             ) {
               Icon(
                 painter = painterResource(R.drawable.ic_add_48dp),
@@ -110,6 +119,8 @@ class PGPKeyListActivity : AppCompatActivity() {
         ) { paddingValues ->
           KeyList(
             identifiers = viewModel.keys,
+            hasSecretKey = ::hasSecretKey,
+            onChangePassphraseClick = ::changeKeyPassphrase,
             onDeleteItemClick = viewModel::deleteKey,
             onExportItemClick = ::exportKey,
             onExportPublicClick = ::exportPublicKey,
@@ -117,7 +128,7 @@ class PGPKeyListActivity : AppCompatActivity() {
             onKeySelected =
               if (isSelecting) {
                 { identifier ->
-                  val keyId = runBlocking { // ensure numeric key ID
+                  val keyId = run { // ensure numeric key ID
                     val key = pgpKeyManager.getKeyById(identifier).unwrap()
                     pgpKeyManager.getKeyId(key) ?: throw NullPointerException()
                   }
@@ -131,6 +142,15 @@ class PGPKeyListActivity : AppCompatActivity() {
         }
       }
     }
+  }
+
+  private fun hasSecretKey(identifier: PGPIdentifier): Boolean =
+    cryptoRepository.hasSecretKey(identifier)
+
+  private fun changeKeyPassphrase(identifier: PGPIdentifier) {
+    val intent = Intent(this, PGPKeyChangePassphraseActivity::class.java)
+    intent.putExtra(PGPKeyChangePassphraseActivity.EXTRA_SELECTED_IDENTIFIER, identifier.toString())
+    keyAction.launch(intent)
   }
 
   private fun exportKey(identifier: PGPIdentifier) {
@@ -208,7 +228,7 @@ class PGPKeyListActivity : AppCompatActivity() {
   }
 
   private fun writeBackupFile(identifier: PGPIdentifier, code: String? = null) {
-    val keyIdAndContent = runBlocking {
+    val keyIdAndContent = run {
       val key = pgpKeyManager.getKeyById(identifier).unwrap()
       val contents =
         if (code != null) { // encrypt secret keys symmetrically
@@ -255,6 +275,11 @@ class PGPKeyListActivity : AppCompatActivity() {
 
     const val EXTRA_SELECTED_KEY = "SELECTED_KEY"
     const val EXTRA_KEY_SELECTION = "KEY_SELECTION_MODE"
+
+    const val PGP_KEY_ADD_REQUEST_KEY = "add_pgp_key"
+    const val ACTION_KEY = "action"
+    const val ACTION_IMPORT_FILE = "from_file"
+    const val ACTION_NEW_PGP_KEY = "generate_new"
 
     fun newSelectionActivity(context: Context): Intent {
       val intent = Intent(context, PGPKeyListActivity::class.java)