+++ crypto/pgpainless/src/main/kotlin/app/passwordstore/crypto/PGPainlessCryptoHandler.kt
@@ -34,7 +34,7 @@ import org.pgpainless.util.Passphrase
 public class PGPainlessCryptoHandler @Inject constructor() :
   CryptoHandler<PGPKey, PGPEncryptOptions, PGPDecryptOptions> {
 
-  public override fun passphraseIsCorrect(key: PGPKey, passphrase: CharArray): Boolean {
+  public override fun passphraseIsCorrect(key: PGPKey, passphrase: CharArray?): Boolean {
     val ciphertextStream = ByteArrayOutputStream()
     val encryptRes =
       encrypt(